# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/jwt/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "JWT login",
                "description": "Login a user and generate JWT token",
                "operationId": "jwtLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "ihamzehald@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "larapoints123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Token type",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration in miliseconds",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGc...",
                                        "token_type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/checkConnection": {
            "get": {
                "tags": [
                    "Check Connection"
                ],
                "summary": "Check internet connection",
                "description": "Fetches all the Asset records",
                "operationId": "d7b799f40b73021de5237079b2bf4f5a",
                "responses": {
                    "200": {
                        "description": "API connection successful!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errcode": {
                                            "description": "The response code",
                                            "type": "integer"
                                        },
                                        "errmsg": {
                                            "description": "The response message",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "errcode": 1,
                                        "errmsg": "ok",
                                        "data": []
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "fail"
                    }
                }
            }
        }
    }
}